+++ src/__tests__/domains/sync/WbsSyncService.test.ts
@@ -2,11 +2,11 @@
 // 以下のテストはWbsSyncServiceの実装変更に伴い、コメントアウトされています。
 
 // import { WbsSyncService } from '@/domains/sync/WbsSyncService';
-// import { ExcelWbs, SyncChanges, SyncError, SyncErrorType } from '@/domains/sync/ExcelWbs';
+// import { ExcelWbs, SyncChanges, SyncError, SyncErrorType } from '@/domains/sync/excel-wbs';
 // import { Phase } from '@/domains/phase/phase';
 // import { PhaseCode } from '@/domains/phase/phase-code';
-// import { IExcelWbsRepository } from '@/applications/wbs-sync/IExcelWbsRepository';
-// import { ISyncLogRepository } from '@/applications/wbs-sync/ISyncLogRepository';
+// import { IExcelWbsRepository } from '@/applications/wbs-sync/iexcel-wbs-repository';
+// import { ISyncLogRepository } from '@/applications/wbs-sync/isync-log-repository';
 // import { IPhaseRepository } from '@/applications/task/iphase-repository';
 // import { IUserRepository } from '@/applications/user/iuser-repository';
 // import { WbsAssignee } from '@/domains/wbs/wbs-assignee';
