+++ src/applications/import-job/import-job-application.service.ts
@@ -4,9 +4,9 @@ import type { IImportJobRepository } from './iimport-job.repository'
 import { ImportJob, ImportJobOptions, ImportJobProgress } from '@/domains/import-job/import-job'
 import { ImportJobStatuses } from '@/domains/import-job/import-job-enums'
 import type { IGeppoImportApplicationService } from '@/applications/geppo-import/geppo-import-application-service'
-import type { IWbsSyncApplicationService } from '@/applications/wbs-sync/IWbsSyncApplicationService'
+import type { IWbsSyncApplicationService } from '@/applications/wbs-sync/iwbs-sync-application-service'
 import { resolveWbsSyncMode } from '@/applications/wbs-sync/wbs-sync-mode'
-import type { INotificationService } from '@/applications/notification/INotificationService'
+import type { INotificationService } from '@/applications/notification/inotification-service'
 import { NotificationType } from '@/types/notification'
 import { NotificationPriority } from '@/domains/notification/notification-priority'
 import { NotificationChannel } from '@/domains/notification/notification-channel'
