+++ src/applications/wbs-sync/iexcel-wbs-repository.ts
@@ -1,4 +1,4 @@
-import { ExcelWbs } from '@/domains/sync/ExcelWbs';
+import { ExcelWbs } from '@/domains/sync/excel-wbs';
 
 export interface IExcelWbsRepository {
   findByWbsName(wbsName: string): Promise<ExcelWbs[] | null>;
